fixed some erros and and debugginfg statements

+++ headers/kernels.h
@@ -3,10 +3,11 @@
 #define KERNELS_H
 
 #include <cuda_runtime.h>
+#include <sys/types.h>
 #include "types.h"
 
 
-__global__ void conflictMatrixKernel(ushort *conflictMatrix, ushort *adjMatrix, ushort *chromosome, unsigned int n); 
+__global__ void conflictMatrixKernel(ushort *conflictMatrix, ushort *adjMatrix, ushort *chromosome, ushort n); 
 std::vector<std::vector<int>*> *crossover(std::vector<int> *first, std::vector<int> *second, int n);
